Setup the basics
This is all a bit disjointed at the moment, but over time this should
come together to make a coherent minecraft server setup for GeekZone.

This uses AWS CDK (Cloud Development Kit) to provision an EC2 and the
required security group in AWS using CloudFormation.

All setup of the minecraft server on the EC2 has been manual so far,
but I fully intend to automate that in this repository.

For now a rough set of manual steps has been added to the main README

I have opened up port 25565 for minecraft and port 22 for SSH. I have
added a `carwyn.pem` for now, but I want to get a proper set of SSH
keys setup fairly soon so that others can access the server.

Notes:

I have based this all on Ubuntu 20.04 since I imagine that is the
version of linux that most are familiar with.

I have written the CDK configuration in Python since a few people
within the GeekZone community seem comofortable with that language.