Setup the minecraft server on boot
This uses the EC2 userdata feature to download and configure spigot on
boot. The downside of the approach is that it takes a while for an EC2
to complete setup. This can be sped up by generating an AMI of the
server which might be a future improvement.

I have set this up to download the world data from S3 which should be
a good basis for backups and whole-system-rebuilds.

The S3 bucket is in a separate stack to avoid new buckets being
created when we update the minecraft-server stack. This means we don't
have to shuffle data between buckets. Also cloudformation does not
delete S3 buckets when the stack is deleted to protect data.

I am using systemd to run the minecraft server. This is more robust
than running a shell script in a screen by hand, but could certainly
be improved with some custom scripting, plugins, or an off the shelf
game server control panel.